+++ src/xstate/actors/audioPlayer/audioPlayerMachine.ts
@@ -10,124 +10,21 @@ import { pure, stop } from 'xstate/lib/actions';
 import { createRadioMachine } from '../radio/radioMachine';
 import { createRepeatMachine } from '../repeatMachine/repeatMachine';
 
+import {
+  getActiveAyahNumber,
+  getActiveVerseTiming,
+  getActiveWordLocation,
+  executeFetchReciter,
+  executeFetchReciterFromEvent,
+  getRecitersList,
+  getMediaSessionMetaData,
+} from './audioPlayerMachineHelper';
 import AudioPlayerContext from './types/AudioPlayerContext';
 import AudioPlayerEventType from './types/AudioPlayerEventType';
 
-import { getAvailableReciters, getChapterAudioData } from 'src/api';
 import { StationType } from 'src/components/Radio/types';
+import { milliSecondsToSeconds } from 'src/utils/datetime';
 import AudioData from 'types/AudioData';
-import Reciter from 'types/Reciter';
-import VerseTiming from 'types/VerseTiming';
-import Word from 'types/Word';
-
-/**
- * check if currentTime is within range timestampFrom and timestampTo
- *
- * example:
- * - timestampFrom = 10, timestampTo = 20, currentTime = 10 should return true
- * - timestampFrom = 10, timestampTo = 20, currentTime = 11 should return true
- * - timestampFrom = 10, timestampTo = 20, currentTime = 20 should return false
- *
- * @param {number} currentTime
- * @param {number} timestampFrom
- * @param {number} timestampTo
- * @returns {boolean} isWithinRange
- */
-const isCurrentTimeInRange = (currentTime: number, timestampFrom: number, timestampTo: number) =>
-  currentTime >= timestampFrom && currentTime < timestampTo;
-
-const getActiveVerseTiming = (context) => {
-  const {
-    audioData: { verseTimings },
-    ayahNumber,
-  } = context;
-  const { currentTime } = context.audioPlayer;
-  const currentTimeMS = currentTime * 1000;
-  const lastAyahOfSurahTimestampTo = verseTimings[verseTimings.length - 1].timestampTo;
-
-  // if the reported time exceeded the maximum timestamp of the Surah from BE, just return the current Ayah which should be the last
-  if (currentTimeMS > lastAyahOfSurahTimestampTo) {
-    return verseTimings[ayahNumber - 1];
-  }
-
-  const activeVerseTiming = verseTimings.find((ayah) => {
-    const isAyahBeingRecited = isCurrentTimeInRange(
-      currentTimeMS,
-      ayah.timestampFrom,
-      ayah.timestampTo,
-    );
-    return isAyahBeingRecited;
-  });
-
-  return activeVerseTiming;
-};
-
-export const getActiveWordLocation = (activeVerseTiming: VerseTiming, currentTime: number) => {
-  const activeAudioSegment = activeVerseTiming.segments.find((segment) => {
-    const [, timestampFrom, timestampTo] = segment; // the structure of the segment is: [wordLocation, timestampFrom, timestampTo]
-    return isCurrentTimeInRange(currentTime, timestampFrom, timestampTo);
-  });
-
-  const wordLocation = activeAudioSegment ? activeAudioSegment[0] : 0;
-  return wordLocation;
-};
-
-const getTimingSegment = (verseTiming: VerseTiming, wordPosition: number) =>
-  verseTiming.segments.find(([location]) => wordPosition === location);
-
-export const getWordTimeSegment = (verseTimings: VerseTiming[], word: Word) => {
-  const verseTiming = verseTimings.find((timing) => timing.verseKey === word.verseKey);
-  if (!verseTiming) return null;
-  const segment = getTimingSegment(verseTiming, word.position);
-  if (segment) return [segment[1], segment[2]];
-  return null;
-};
-
-const getActiveAyahNumber = (activeVerseTiming: VerseTiming) => {
-  const [, verseNumber] = activeVerseTiming.verseKey.split(':');
-  return Number(verseNumber);
-};
-
-const executeFetchReciter = async (context: AudioPlayerContext): Promise<AudioData> => {
-  const { reciterId, surah } = context;
-  return getChapterAudioData(reciterId, surah, true);
-};
-
-const executeFetchReciterFromEvent = async (
-  context: AudioPlayerContext,
-  event,
-): Promise<AudioData> => {
-  const { surah } = event;
-  const { reciterId } = context;
-  // @ts-ignore
-  const data = await executeFetchReciter({ reciterId, surah });
-  return {
-    ...data,
-    ...event,
-  };
-};
-
-const getMediaSessionMetaData = async (context: AudioPlayerContext, recitersList: Reciter[]) => {
-  const reciterName = recitersList.find((reciter) => reciter.id === context.reciterId).name;
-  return new MediaMetadata({
-    title: `Surah ${context.audioData.chapterId}`,
-    artist: reciterName,
-    album: 'Quran.com',
-    artwork: [
-      {
-        src: 'https://quran.com/images/logo/Logo@192x192.png',
-        type: 'image/png',
-        sizes: '192x192',
-      },
-    ],
-  });
-};
-
-const getRecitersList = async (context: AudioPlayerContext) => {
-  const { recitersList } = context;
-  if (recitersList) return recitersList;
-  return getAvailableReciters('en').then((res) => res.reciters);
-};
 
 export const audioPlayerMachine =
   /** @xstate-layout N4IgpgJg5mDOIC5QEMCuECWB7ACgG2QE8wAnAOgAkBJAERoFEA5AYhwBkBBATQH0AlDjSoB5RKAAOWWBgAu2AHZiQAD0QAOAJxkADJo0A2DWu0B2AIxqTAZgBMAGhCFEAVltkALBq-7DJ5140rAF8ghzRMXAJicmo6JlZOXgEhYR4AFQEAYQBpJUlpOSxFJBVEK20zMhNNGzV9GzN3bWcLBycEG2sPAM99E20GmytnELD0bHwiUkpaBhZ2bh5uDgo8qVkFJVUEcsrqjVr6xubWx0R3KzVur2H9Y2MbZ3dRkHCJqOnYuYTFgGUAVQEqxK+Q2RS2ZQqVRqdQaTRaajaiH0Zkqni87mM7gad2eoVe40iUxis3iv3oaR4Cy4ACEODkeL8cPR6DQ1gVNiVtmZtFYyGp-O4fEN9M41FYDEidmoruiNCYbF5nNptEKTC83kTojM4ixMhQOIwAOL0fj0TJUNL0PjssHFUDcixkMxWXH+ZV+Mz+KU2dzua4GEw3cVeDWEybagBqVF+VBpbHoZABQJ4AFkY6mOGl9cxMsJGAAxKh8VNUxJpjNZ-WsxmAla2wr20oIVGi508-qGMz6Wx1KUtCVVF0yqyuoXisMRCPTaOx+OJ5MrCu-TPZii5w2ZehsMuLdMrqsUGuL4ESdaNiEtizOdumbRdnuHKVex5kALmKx+J4KtST97EshZzjBMyA4f4Ul3LhrR4KhGEtKgs1ZMgcDA8kaFAzI0ioSN6GYNJhCNI0EwbTkHUhfRtDIRUFU8MwNG7FUTB9GUTCo45nBsQY+ncdV8U1adyCA+dQPAkRIOg2D4MQ9CUP+NCMKwnDmHJehchBc9SObR5zB0ao-B4lFDF9H0-T2F1+gefplTxMYpw+QSY2AxMwIg6kJLgrDpOQ1CkPpRTcN+NIODYNgSPBLlEG0ypTEsb9DIOdwfUMf0P3MDR3DMGxtA0Zo-y1GdHOElyxLcvgYI8hCrRknz0L87DcOpMKm22coKKo+VfXo+iKNMKVyi6BpLiMCpdDo4I+PDezAMKkDitSUryqkqrvLk3zMPq5gmDZdSOXCsidlVSjqM6uiGN6s4dkCG9UvFXozGqfQ8oE6a51m0T5sSdylqQ2T5LYYRBFgo1mEYegAA1KWWU8QFBC8Io6awbwFLK1B49xnCDEy7iHT9UYeTKnqmoS3tcz6yskzzlt+pD-sB41WD4ehIyWLh6x2u1L06YZ+Q43Q0YxjQfSy1iP1FTRVXS0dCYA4nnPe8TyYqrzqfQ2mhHpgsOCoUL2bh-b0fo517mcayeKGZ8Whvd9MRldLON-Ca7Jlma5dJ7gvspn6arINWgY3RhIKazmKJsMhXAMUVamol0fQlPkP00YwXVsIZpajF2RLdqDFe+9CGESP3QYhlm2bPXbmshYx2porqzqYi77pla4rMVQxmhsNOCte12SrJxbPbz7duD9nBGeZqGg-h8oq+O2jusYkzLDYiV-DUIYcod2z-3T7vM9792c4Hsh8+H+nlFgGRkBkMAyGQAAza+SAAChwvhyR4E+uAASmYfiiYzuaCt+6VSQp-IGk99qtSOh1OeddmKDkGrbE2wxVSdwcrvQBC0KYgJkgXY0CkNr4UIsRXWmkWqfi0CYC4qIYTUT6llLQg1rD3isNiAYaCXpOT3h9A+wDlZ4KNAQpSjNmRZhLtDWGZCygUKqNQiwBw17yj6tif0TDOjNB4mKXiW98roK4Zgvu2D+Gn0EXVJSKk1Jlw5lPUUlQJRGA0MqCUMV67tAaPeHmfQvCmGcD2DhstuFAKMVTARQiApBRChA5sroWhkHsZoJxOU9Kx38J4tQPIhT3m0QSJ2O99HyywUrEJJiwmbUYNtKxetmwZTaiiHK6U-BGAVElV8ApuJZJysg-xACCmGKKT9UJZjcL-BwDQRC6QqDpmNFElqQoUpyNoUoi6jwPG3B8LCRothHqO23l3fJWcPY4OQoM9aSlqYzOkUGWRVgaEKLoRdV0zRl4mFFC8u4SpukYN6bw4JAySm+3pkXSGrMJEaT2tE8U+hnTWE4sqIY90MY+k2W+LwVl6gDBRJ8-Z+9s58OKVwIGPsAbq2BqPJm4iLnSldNClOcKbl+FcYgCwa8UXyl8aqMUzRsl-2dl8g5h8jnUkJQC4GmttaUsyZUCwugTaqjNlYZ8tgrZeFRmKYYDDuWTV5dinhuLfm4P+cSv2mRDSB1IeClqdxQ51F0M0dlwxGUdGaFcNpDj7x0XYTs3RnCirfL1f09CW1WTKRZJYmGYKK4dFYVcAwBgZTZXSfdBVF0mgHBxmvAUHETYaCxb6-leKkJBpoD8LglLRxXKoTc+RUclntE-K4HG+h0amFYd+XNJMcWHK8kWktFLzWRuvFoVULzeydBNj2KUmJfTpruDKROm8cm7L0XmztAru3lODdSWsQJKWDp0AZUdX4J0psuNdYc9j62hi9c9AJBifkBrID2ohRF6ASuOGQOp2VNFNPsA3DirF0RBmsPUG5fjr3-z5augtgaN3Fq3ckMSGR6RhskRaplAoh0Hqjke5N7RXCqhnT4DJJhNW5L2Su3VXaqrMBEfQMRRY4K-CPBU8N5dObdi0J4WoK8m2OPuj6JJ-IDg9lMPKIw2ydE3p6fm-VINwbAtLqx6xkDWofvKIqV0jjXR+ClAcViAR5SjkMmvUjS6fUdso2u6jZLx4grLapkTGnDCuD6M4fsfhWWOO-KYGwYHJMQa4SKs0ForRlTGUFZgEAig3wwPIAAblgAA1jfO+YAZAAGMAAWfAwDpdkKQXdGH90juw+O3DTKfB8nfAOVuHZ22JiC4zEL0FwscGYKQEgWByDiAIDIO+XWAC2ZBUsZey7l-LJBCupOHY+EzZX+zKjMpcZtGNuL1aJXTI0ZpRGUla5F6LZBYsJeS8NtLWWcviDAFfGgV9kC7oqHyC4zqjAZU4ulfsFxPOpsCCRxx63Gv0B2x-LMbWOtdbID1q+-WSBDZG+dsAl3ru3fuy6Dw5QxQvcyoqRKKbuyecMA0vo5QbKLu9bLXM-1ySUqeNa+iVavF9EVFKTQnnqjpMuCR8a-ntXzhDZSakdIGRMhZCx1DkbOiBHap+HibPfPin7OKMyTRm3o14-V3MBpjSmia5aa0lLDJhweq89KfoXkfcMOmjiRwsnq-g4DVISGcj67xxjOoxu-QGX7MTwjKJh2mbJy7XtE9+2XgN67nwAsPdm4bncSovim3WBuba-3Unu69pPM7qF4f3em-0M+HilF48OrXsMQMtvywIdECH+GnRqhh19I8F0YpLC-rw1Qs9lxuINA+eB4kfOljy0ZgWSljfWIJJdIYWwnuLo9TPRlcwcu7gjF79EfvtGxHkjSFhaZ1f9qIobi81nmgXnSoMCEfE8gsAQDgEoHl2ovhMH1635EQwAxWpeYEBo62TzLlXPqSl8olEzKyovI5gPYdwFsJscSqKTwhmnEfo62d6-qucK08kQyI+FwMahgs6CaqI1gAm6Mzo2I46ji942IXOpOqeOqQSD6KsG2JKI+Lyqi8KXoicIBbmyy1gKULo9E94XgrciBfqVGXsq0ouEanMGU-osaOBRgeB5WTqRgOMTBPYziy+3OeSFGNBKBYCO+lSUiHQHG3Q3GjivGLQjqdEyo0BgYlwzQoGFBd+5GFmWhR8Qq+C6Bu+0Sng-oLQvgMSHq94SUUBtwQYmIHuhgghMmtBoSIqZadQ3hfoFEMog0TOyyqarKPELEdEf2K+jhPclm0GxyJiZafolElatyNa5hCooc1sc6ds9hWqGhThhSKB1AgUwgfApaHh2wDQVCRhIYvingZhFslwzchmIG3Y9RZGy6TRfSKBRaEq1COgvo1gMoxBMofUQYVwTCBsnEh0KeAWmhzRA8I+jQN4XG-Rph-GDcTQUKcovI2UmU1QHcOR0xiYYqCYYhbG8MQo8oukXUoon45QgsDcgJzoKqFQAop03Y-2RqxowWuuYWIOu6WaQ4DCKoEokceeF0rgqOCc1QpgQYX+Lx5mDWsJW26+u2SJXRL+ceRukeuek6GUnma87ywsOaxJss1Oaa2e9J0+biji+O34ngo47J6hBW1JOw8hUKKoMpspspjQ5+QQQAA */
@@ -789,7 +686,7 @@ export const audioPlayerMachine =
         }),
         setAudioData: assign({
           duration: (context, event: any) => {
-            return event.data.duration / 1000;
+            return milliSecondsToSeconds(event.data.duration);
           },
           audioData: (context, event: any) => event.data,
           surahVersesCount: (context, event: any) => event.data.verseTimings.length,
@@ -813,7 +710,7 @@ export const audioPlayerMachine =
           } else {
             const ayahTimestamps = verseTimings[ayahNumber - 1];
             const { timestampFrom } = ayahTimestamps;
-            context.audioPlayer.currentTime = timestampFrom / 1000;
+            context.audioPlayer.currentTime = milliSecondsToSeconds(timestampFrom);
           }
         },
         playAudio: (context) => {
